+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [2.0.2] - 2018-12-10
+- More DRY approach for gateways.
+- Added support for payment lines.
+- Added Billink and Capayable gateways.
+
 ## [2.0.1] - 2018-07-06
 - Added fallback to the default Pronamic Pay configuration ID.
 - Prefixed the Pronamic gateways with 'Pronamic - '.
@@ -60,12 +65,13 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 - Tested on Easy Digital Downloads version 2.3.
 
 ## [1.0.1] - 2015-03-03
-- Changed WordPress pay core library requirment from ~1.0.0 to >=1.0.0.
+- Changed WordPress pay core library requirment from `~1.0.0` to `>=1.0.0`.
 
 ## 1.0.0 - 2015-01-20
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/easy-digital-downloads/compare/2.0.1...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/easy-digital-downloads/compare/2.0.2...HEAD
+[2.0.2]: https://github.com/wp-pay-extensions/easy-digital-downloads/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/wp-pay-extensions/easy-digital-downloads/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/wp-pay-extensions/easy-digital-downloads/compare/1.2.7...2.0.0
 [1.2.7]: https://github.com/wp-pay-extensions/easy-digital-downloads/compare/1.2.6...1.2.7